Refactor faq plugin
