+++ src/default-theme/index.css
@@ -26,7 +26,7 @@
 @import '../console/index.css';
 @import '../csvwidget/index.css';
 @import '../editorwidget/index.css';
-@import '../faq/index.css';
+@import '../plugin-faq/index.css';
 @import '../filebrowser/index.css';
 @import '../help/index.css';
 @import '../imagewidget/index.css';
