--- src/faq/index.ts
@@ -1,14 +1,0 @@
-// Copyright (c) Jupyter Development Team.
-// Distributed under the terms of the Modified BSD License.
-
-export * from './widget';
-
-
-/**
- * The command IDs used by the FAQ plugin.
- */
-export
-namespace CommandIDs {
-  export
-  const open: string = 'faq-jupyterlab:open';
-};
