+++ src/plugin-faq/index.css
