+++ src/plugin-faq/widget.ts
